Updated source files
Added resign command
Changed places of thread functions in botChess
Moves bots Python classes to bots folder
Added treatment for when OBS json does not exists
