+++ .gitignore
@@ -1,2 +1,3 @@
 config/config.py
 *.pyc
+.vscode/
